Added static_cast<float>(...) on certain results to prevent integer division truncation errors

+++ tree/decision-tree-classifier/tree.cpp
@@ -46,7 +46,7 @@ vec DecisionTreeClassifier::compute_class_probability (vec& Y) {
     for (int i = 0; i < unq_ret.labels.size(); i++) {
         if (classes_to_index.contains(unq_ret.labels[i])) {
             int index = classes_to_index.at(unq_ret.labels[i]);
-            prob_vec.at(index) = static_cast<float>(unq_ret.label_counts[i]) / Y.n_rows;
+            prob_vec.at(index) = static_cast<float>(unq_ret.label_counts[i]) / static_cast<float>(Y.n_rows);
         }
     }
 
@@ -80,8 +80,8 @@ float DecisionTreeClassifier::compute_information_gain (
     float left_subset_imp = DecisionTreeClassifier::determine_impurity_metric(left_subset);
     float right_subset_imp = DecisionTreeClassifier::determine_impurity_metric(right_subset);
 
-    float left_subset_weighted = (static_cast<float>(left_subset.n_rows) / Y.n_rows) * left_subset_imp;
-    float right_subset_weighted = (static_cast<float>(right_subset.n_rows) / Y.n_rows) * right_subset_imp;
+    float left_subset_weighted = (static_cast<float>(left_subset.n_rows) / static_cast<float>(Y.n_rows)) * left_subset_imp;
+    float right_subset_weighted = (static_cast<float>(right_subset.n_rows) / static_cast<float>(Y.n_rows)) * right_subset_imp;
 
     return main_data_imp - (left_subset_weighted + right_subset_weighted);
 }
